+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
+## v0.0.13
+  * Mark metrics with non-ascii non-alphanumeric characters as unsupported [#28](https://github.com/singer-io/tap-ga4/pull/28)
+## v0.0.12
+  * Update premade report names to match GA4 Data API changes [#27](https://github.com/singer-io/tap-ga4/pull/27)
 ## v0.0.11
-  * Update field names to match GA4 Data API changes  [#26](https://github.com/singer-io/tap-ga4/pull/26)
+  * Update field names to match GA4 Data API changes [#26](https://github.com/singer-io/tap-ga4/pull/26)
 ## v0.0.10
   * Update datetime formats to account for canonicalization of datetime dimensions  [#22](https://github.com/singer-io/tap-ga4/pull/22)
 ## v0.0.9